requirements.txt
deepface
matplotlib
opencv-python
pandas
Pillow
pytube
scipy
scikit-learn
seaborn
tqdm
wbgapi
basemap
unidecode
xlrd
gTTS
pydub
python-dotenv
PyVimeo
eurostat
moviepy
python-forecastio
git+https://github.com/PrithivirajDamodaran/Parrot_Paraphraser.git

